requirements.txt
playwright~=1.21.0
Faker~=13.4.0
smsactivate~=1.3
mysql-connector-python~=8.0.29
requests~=2.27.1